Desktop/RAG Data Pipeline/requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Machine Learning
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-cloud-storage>=2.10.0

# Vector Operations
numpy>=1.24.3
faiss-cpu>=1.7.4

# HTTP Clients
httpx>=0.25.0
requests>=2.31.0

# Data Processing
tiktoken>=0.5.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Evaluation
ragas>=0.0.18
langsmith>=0.0.69

# Utilities
python-json-logger>=2.0.7
structlog>=23.2.0
tenacity>=8.2.3

# Development
black>=23.9.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.6.0

# Production
gunicorn>=21.2.0